fix: split large vector database file
